+++ lib/core/constants/enums.dart
@@ -31,3 +31,12 @@ enum PetType {
   final String displayName;
   final String code;
 }
+
+enum HomeHomeTab {
+  petSos('돌봄급구'),
+  petMate('돌봄메이트');
+
+  const HomeHomeTab(this.label);
+
+  final String label;
+}
