+++ lib/presentation/pages/home/providers/show_search_bar_provider.dart
@@ -1,0 +1,13 @@
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'show_search_bar_provider.g.dart';
+
+@riverpod
+class ShowSearchBar extends _$ShowSearchBar {
+  @override
+  bool build() {
+    return false;
+  }
+
+  set show(bool show) => state = show;
+}
